Remove unused java classes to improve test coverage
