--- janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/ReadOnlyKeyColumnValueStore.java
@@ -1,45 +1,0 @@
-// Copyright 2017 JanusGraph Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.janusgraph.diskstorage.keycolumnvalue;
-
-import org.janusgraph.diskstorage.BackendException;
-import org.janusgraph.diskstorage.Entry;
-import org.janusgraph.diskstorage.StaticBuffer;
-
-import java.util.List;
-
-/**
- * Wraps a {@link KeyColumnValueStore} and throws exceptions when a mutation is attempted.
- *
- * @author Matthias Br&ouml;cheler (me@matthiasb.com);
- */
-public class ReadOnlyKeyColumnValueStore extends KCVSProxy {
-
-    public ReadOnlyKeyColumnValueStore(KeyColumnValueStore store) {
-        super(store);
-    }
-
-    @Override
-    public void acquireLock(StaticBuffer key, StaticBuffer column, StaticBuffer expectedValue,
-                            StoreTransaction txh) throws BackendException {
-        throw new UnsupportedOperationException("Cannot lock on a read-only store");
-    }
-
-    @Override
-    public void mutate(StaticBuffer key, List<Entry> additions, List<StaticBuffer> deletions, StoreTransaction txh) throws BackendException {
-        throw new UnsupportedOperationException("Cannot mutate a read-only store");
-    }
-
-}
