--- janusgraph-core/src/main/java/org/janusgraph/util/datastructures/IntSet.java
@@ -1,30 +1,0 @@
-// Copyright 2017 JanusGraph Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.janusgraph.util.datastructures;
-
-
-public interface IntSet {
-
-    boolean add(int value);
-
-    boolean addAll(int[] values);
-
-    boolean contains(int value);
-
-    int[] getAll();
-
-    int size();
-
-}
